Fix alert positioning on detail pages
