+++ context/app/static/js/components/detailPage/DetailLayout/DetailLayout.jsx
@@ -1,43 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import useEntityStore, { savedAlertStatus, editedAlertStatus } from 'js/stores/useEntityStore';
 import TableOfContents from 'js/shared-styles/sections/TableOfContents';
 import { getSections } from 'js/shared-styles/sections/TableOfContents/utils';
-import { Content, FlexRow, StyledAlert } from './style';
+import SharedAlerts from 'js/components/detailPage/SharedAlerts';
+import { Content, FlexRow, StyledDiv } from './style';
 
-const entityStoreSelector = (state) => ({
-  shouldDisplaySavedOrEditedAlert: state.shouldDisplaySavedOrEditedAlert,
-  setShouldDisplaySavedOrEditedAlert: state.setShouldDisplaySavedOrEditedAlert,
-});
-
-function DetailLayout(props) {
-  const { sectionOrder, children } = props;
-
-  const { shouldDisplaySavedOrEditedAlert, setShouldDisplaySavedOrEditedAlert } = useEntityStore(entityStoreSelector);
-
+function DetailLayout({ sectionOrder, additionalAlerts, children }) {
   // section hash must match section id in each component
   const sections = new Map(getSections(sectionOrder));
 
   return (
-    <>
-      {shouldDisplaySavedOrEditedAlert === savedAlertStatus && (
-        <StyledAlert severity="success" onClose={() => setShouldDisplaySavedOrEditedAlert(false)}>
-          Successfully added to My Saves List. All lists are currently stored on local storage and are not transferable
-          between devices.
-        </StyledAlert>
-      )}
-      {shouldDisplaySavedOrEditedAlert === editedAlertStatus && (
-        <StyledAlert severity="success" onClose={() => setShouldDisplaySavedOrEditedAlert(false)}>
-          Successfully updated save status. All lists are currently stored on local storage and are not transferable
-          between devices.
-        </StyledAlert>
-      )}
+    <StyledDiv>
+      <SharedAlerts additionalAlerts={additionalAlerts} />
       <FlexRow>
         <TableOfContents items={[...sections.values()]} />
         <Content>{children}</Content>
       </FlexRow>
-    </>
+    </StyledDiv>
   );
 }
