+++ context/app/static/js/components/detailPage/SharedAlerts/index.js
@@ -1,0 +1,3 @@
+import SharedAlerts from './SharedAlerts';
+
+export default SharedAlerts;
