+++ context/app/static/js/components/detailPage/SupportAlert/SupportAlert.jsx
@@ -3,12 +3,12 @@ import React from 'react';
 import { LightBlueLink } from 'js/shared-styles/Links';
 import Typography from '@material-ui/core/Typography';
 
-import { DetailPageAlert } from 'js/components/detailPage/style';
+import { Alert } from 'js/shared-styles/alerts';
 
 function SupportAlert({ uuid }) {
   // There should usually be only one parent, but this is more robust, and we want to keep it simple.
   return (
-    <DetailPageAlert severity="warning">
+    <Alert severity="warning">
       <Typography variant="body2">
         “Support” entities provide derived, low-level data for visualizations. Navigate to{' '}
         <LightBlueLink href={`/search?descendant_ids[0]=${uuid}&entity_type[0]=Dataset`}>
@@ -16,7 +16,7 @@ function SupportAlert({ uuid }) {
         </LightBlueLink>{' '}
         for a view of this information in context.
       </Typography>
-    </DetailPageAlert>
+    </Alert>
   );
 }
