+++ context/app/static/js/pages/Dataset/Dataset.jsx
@@ -14,9 +14,9 @@ import ContributorsTable from 'js/components/detailPage/ContributorsTable';
 import useSendUUIDEvent from 'js/components/detailPage/useSendUUIDEvent';
 import CollectionsSection from 'js/components/detailPage/CollectionsSection';
 import SupportAlert from 'js/components/detailPage/SupportAlert';
-import { DetailPageAlert } from 'js/components/detailPage/style';
 import { useSearchHits } from 'js/hooks/useSearchData';
 import { getAllCollectionsQuery } from 'js/helpers/queries';
+import { Alert } from 'js/shared-styles/alerts';
 
 // TODO use this context for components other than FileBrowser
 import DetailContext from 'js/components/detailPage/context';
@@ -75,8 +75,8 @@ function DatasetDetail(props) {
     registered_doi,
     doi_url,
     contributors,
+    next_revision_uuid,
   } = assayMetadata;
-  const isLatest = !('next_revision_uuid' in assayMetadata);
 
   const combinedStatus = sub_status || status;
 
@@ -115,22 +115,28 @@ function DatasetDetail(props) {
   // TODO: When all environments are clean, data_types array fallbacks shouldn't be needed.
   return (
     <DetailContext.Provider value={{ hubmap_id, uuid, mapped_data_access_level }}>
-      {!isLatest && (
-        <DetailPageAlert severity="warning" $marginBottom="16">
-          <span>
-            {/* <span> to override "display: flex" which splits this on to multiple lines. */}
-            You are viewing an older version of this page. Navigate to the{' '}
-            <LightBlueLink href={`/browse/latest/dataset/${uuid}`}>latest version</LightBlueLink>.
-          </span>
-        </DetailPageAlert>
-      )}
-      {mapped_external_group_name && (
-        <DetailPageAlert severity="info" $marginBottom="16">
-          You are viewing an external dataset that was not generated by the HuBMAP Consortium.
-        </DetailPageAlert>
-      )}
-      {entity_type === 'Support' && <SupportAlert uuid={uuid} />}
-      <DetailLayout sectionOrder={sectionOrder}>
+      <DetailLayout
+        sectionOrder={sectionOrder}
+        additionalAlerts={
+          <>
+            {next_revision_uuid && (
+              <Alert severity="warning" $marginBottom="16">
+                <span>
+                  {/* <span> to override "display: flex" which splits this on to multiple lines. */}
+                  You are viewing an older version of this page. Navigate to the{' '}
+                  <LightBlueLink href={`/browse/latest/dataset/${uuid}`}>latest version</LightBlueLink>.
+                </span>
+              </Alert>
+            )}
+            {['Dataset', 'Support'].includes(mapped_external_group_name) && (
+              <Alert severity="info" $marginBottom="16">
+                You are viewing an external dataset that was not generated by the HuBMAP Consortium.
+              </Alert>
+            )}
+            {entity_type === 'Support' && <SupportAlert uuid={uuid} $marginBottom="16" />}
+          </>
+        }
+      >
         <Summary
           uuid={uuid}
           entity_type={entity_type}
